requirements.txt
werkzeug
psycopg2-binary
flask
gunicorn
google-genai
requests
posthog
dotenv
firebase-admin

